refactor the project for further development
