+++ app/core/config.py
@@ -1,0 +1,1 @@
+# load env vars
