+++ app/crud/job_crud.py
@@ -1,8 +1,8 @@
 # Contains DB logic.
 
 from sqlalchemy.orm import Session
-import app.models as models
-import app.schemas as schemas
+import app.models.job_models as models
+import app.schemas.job_schemas as schemas
 from fastapi import HTTPException
 import logging
