+++ app/models/job_models.py
@@ -3,7 +3,7 @@
 # defining how Python classes = database tables.
 
 from sqlalchemy import Column, Integer, String, Date
-from app.database import Base
+from app.db.database import Base
 
 
 class JobApplication(Base):
