+++ tests/conftest.py
@@ -2,7 +2,7 @@
 import pytest
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-from app.database import Base
+from app.db.database import Base
 from app.main import app
 from fastapi.testclient import TestClient
